windows_agent/requirements.txt
ultralytics>=8.2.0
onnxruntime>=1.18.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
requests>=2.31.0
paho-mqtt>=1.6.1
pystray>=0.19.5
Pillow>=10.3.0
pyinstaller>=6.6.0
psutil>=5.9.8
schedule>=1.2.2
